Search objects pattern and exact alias

Approved-by: Idan Dunsky

+++ src/main/java/aii/logic/CommandsServiceImplementation.java
@@ -1,8 +1,8 @@
 package aii.logic;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Value;
@@ -13,9 +13,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import aii.dal.CommandsCrud;
 import aii.data.CommandEntity;
+import aii.data.UserRole;
 import aii.logic.exceptions.InvalidCommandException;
 import aii.logic.exceptions.InvalidInputException;
+import aii.logic.exceptions.UserUnauthorizedException;
 import aii.logic.utilities.EmailValidator;
+
+
 
 
 @Service
@@ -23,11 +27,15 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
     private CommandsCrud commands;
     private String springApplicationName;
     private EmailValidator emailValidator;
+    private EnhancedUsersService users;
+    private EnhancedObjectsService objects;
 
 
 
-    public CommandsServiceImplementation(CommandsCrud commands){
+    public CommandsServiceImplementation(CommandsCrud commands, EnhancedUsersService users, EnhancedObjectsService objects) {
         this.commands = commands;
+        this.users = users;
+        this.objects = objects;
         emailValidator = new EmailValidator();
     }
 
@@ -109,8 +117,22 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
             throw new InvalidCommandException("ERROR - UserId - SystemID is empty");
         }
 
-        CommandEntity commandEntity = new CommandEntity();
+        // Validate user's permission:
+        UserRole role = this.users.getUserRole(newCommand.getInvokedBy().getUserId().getSystemID(),
+                newCommand.getInvokedBy().getUserId().getEmail());
 
+        if (role != UserRole.END_USER) {
+            throw new InvalidCommandException("ERROR - User does not have permission to invoke commands");
+        }
+
+        ObjectId obj = newCommand.getTargetObject().getObjectId();
+        UserId ui = newCommand.getInvokedBy().getUserId();
+
+        // Check for object existence and active is true (this method will throw if the conditions are not met.)
+        Optional<ObjectBoundary> op = this.objects.getSpecificObject(ui.getSystemID(), ui.getEmail(), obj.getSystemID(), obj.getId());
+
+        // Passed all validations, create and save the command entity:
+        CommandEntity commandEntity = new CommandEntity();
         CommandId commandId = new CommandId(this.springApplicationName, UUID.randomUUID().toString());
         commandEntity.setCommandId(commandId);
 
@@ -133,6 +155,8 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
     @Override
     @Deprecated
     public List<CommandBoundary> getAllCommands(String adminSystemID, String adminEmail) {
+        throw new UnsupportedOperationException("ERROR - Deprecated method");
+        /*
         // Validate admin credentials
         if (adminSystemID == null || adminEmail == null) {
             throw new InvalidInputException("ERROR - Admin credentials are required");
@@ -142,6 +166,11 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
             throw new InvalidInputException("ERROR - Invalid email format");
         }
 
+        UserRole role = this.users.getUserRole(adminSystemID, adminEmail);
+
+        if (role != UserRole.ADMIN) {
+            throw new UserUnauthorizedException("ERROR - User does not have permission to view all commands");
+        }
 
         // Fetch all command entities from the DB
         List<CommandEntity> commandEntities = this.commands.findAll();
@@ -153,6 +182,7 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
         }
 
         return commandBoundaries;
+        */
     }
 
     @Override
@@ -165,6 +195,12 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
 
         if (!emailValidator.isEmailValid(adminEmail)) {
             throw new InvalidInputException("ERROR - Invalid email format");
+        }
+
+        UserRole role = this.users.getUserRole(adminSystemID, adminEmail);
+
+        if (role != UserRole.ADMIN) {
+            throw new UserUnauthorizedException("ERROR - User does not have permission to view all commands");
         }
 
         return this.commands
@@ -183,12 +219,15 @@ public class CommandsServiceImplementation implements EnhancedCommandService {
             throw new InvalidInputException("[ERROR] - Admin credentials are required");
         }
 
-        // TODO: Validate admin credentials
-
         if (!emailValidator.isEmailValid(adminEmail)) {
             throw new InvalidInputException("[ERROR] - Invalid email format");
         }
 
+        UserRole role = this.users.getUserRole(adminSystemID, adminEmail);
+
+        if (role != UserRole.ADMIN) {
+            throw new UserUnauthorizedException("[ERROR] - User does not have permission to delete all commands");
+        }
 
         this.commands.deleteAll();
         System.out.println("[WARN] - All commands deleted by admin: " + adminSystemID + " / " + adminEmail);
